Fixed: Weapon of the sorcerer
Fixed a bug with Archmage Proudmoore's skill Weapon of the Sorcerer which caused the player to drop the weapons on the ground if he already had weapons equipped.
The skill will now remove the weapons the player is holding and give him the two new conjured ones as originally intended.